优化代码，修复服务器无法关闭问题

+++ Include/mysysinfo/mysysinfo.cpp
@@ -29,7 +29,7 @@ bool MySysInfo::GetMemUsage(double& nMemTotal, double& nMemUsed) {
 	auto lst = str.split(" ");
 	if (lst.size() > 6) {
 		nMemTotal = lst[1].toDouble();
-		nMemUsed  = nMemTotal - lst[6].toDouble();
+		nMemUsed = nMemTotal - lst[6].toDouble();
 		return true;
 	}
 	else {
@@ -43,8 +43,8 @@ bool MySysInfo::GetMemUsage(double& nMemTotal, double& nMemUsed) {
 		return false;
 	}
 	double nMemFree = memsStat.ullAvailPhys / (1024.0 * 1024.0);
-	nMemTotal		= memsStat.ullTotalPhys / (1024.0 * 1024.0);
-	nMemUsed		= nMemTotal - nMemFree;
+	nMemTotal = memsStat.ullTotalPhys / (1024.0 * 1024.0);
+	nMemUsed = nMemTotal - nMemFree;
 	return true;
 #endif
 }
@@ -78,7 +78,7 @@ bool MySysInfo::GetNetUsage() {
 #if defined(Q_OS_WIN32)
 __int64 Filetime2Int64(const FILETIME* ftime) {
 	LARGE_INTEGER li;
-	li.LowPart	= ftime->dwLowDateTime;
+	li.LowPart = ftime->dwLowDateTime;
 	li.HighPart = ftime->dwHighDateTime;
 	return li.QuadPart;
 }
@@ -98,15 +98,15 @@ bool MySysInfo::GetCpuUsage(double& nCpuRate) {
 	str.replace(QRegExp("( ){1,}"), " ");
 	auto lst = str.split(" ");
 	if (lst.size() > 3) {
-		double use	 = lst[1].toDouble() + lst[2].toDouble() + lst[3].toDouble();
+		double use = lst[1].toDouble() + lst[2].toDouble() + lst[3].toDouble();
 		double total = 0;
 		for (int i = 1; i < lst.size(); ++i)
 			total += lst[i].toDouble();
 		if (total - m_cpu_total__ > 0) {
 			// qDebug("cpu rate:%.2lf%%", (use - m_cpu_use__) / (total - m_cpu_total__) * 100.0);
 			m_cpu_total__ = total;
-			m_cpu_use__	  = use;
-			nCpuRate	  = (use - m_cpu_use__) / (total - m_cpu_total__) * 100.0;
+			m_cpu_use__ = use;
+			nCpuRate = (use - m_cpu_use__) / (total - m_cpu_total__) * 100.0;
 			return true;
 		}
 	}
@@ -119,17 +119,17 @@ bool MySysInfo::GetCpuUsage(double& nCpuRate) {
 	FILETIME idleTime;
 	FILETIME kernelTime;
 	FILETIME userTime;
-	res			  = GetSystemTimes(&idleTime, &kernelTime, &userTime);
-	preIdleTime	  = idleTime;
+	res = GetSystemTimes(&idleTime, &kernelTime, &userTime);
+	preIdleTime = idleTime;
 	preKernelTime = kernelTime;
-	preUserTime	  = userTime;
-	hEvent		  = CreateEvent(nullptr, FALSE, FALSE, nullptr); //初始值为nonsignaled
-	WaitForSingleObject(hEvent, 500);							 //等待500毫秒
-	res				 = GetSystemTimes(&idleTime, &kernelTime, &userTime);
-	long long idle	 = CompareFileTime(preIdleTime, idleTime);
+	preUserTime = userTime;
+	hEvent = CreateEvent(nullptr, FALSE, FALSE, nullptr); //初始值为nonsignaled
+	WaitForSingleObject(hEvent, 500);					  //等待500毫秒
+	res = GetSystemTimes(&idleTime, &kernelTime, &userTime);
+	long long idle = CompareFileTime(preIdleTime, idleTime);
 	long long kernel = CompareFileTime(preKernelTime, kernelTime);
-	long long user	 = CompareFileTime(preUserTime, userTime);
-	nCpuRate		 = ceil(100.0 * (kernel + user - idle) / (kernel + user));
+	long long user = CompareFileTime(preUserTime, userTime);
+	nCpuRate = ceil(100.0 * (kernel + user - idle) / (kernel + user));
 	// qDebug() << "windows:CPU use rate:" << nCpuRate << "%";
 #endif
 	return true;
@@ -148,7 +148,7 @@ bool MySysInfo::GetDiskSpeed() {
 	auto lst = str.split(" ");
 	if (lst.size() > 5) {
 		// qDebug("disk read:%.0lfkb/s disk write:%.0lfkb/s", (lst[4].toDouble() - m_disk_read__) / (m_timer_interval__ / 1000.0), (lst[5].toDouble() - m_disk_write__) / (m_timer_interval__ / 1000.0));
-		m_disk_read__  = lst[4].toDouble();
+		m_disk_read__ = lst[4].toDouble();
 		m_disk_write__ = lst[5].toDouble();
 		return true;
 	}
@@ -168,7 +168,7 @@ bool MySysInfo::GetdiskSpace(unsigned long& lFreeAll, unsigned long& lTotalAll) 
 			str.replace(QRegExp("( ){1,}"), " ");
 			auto lst = str.split(" ");
 			if (lst.size() > 5)
-				// qDebug("挂载点:%s 已用:%.0lfMB 可用:%.0lfMB", lst[5].toStdString().c_str(), lst[2].toDouble() / 1024.0, lst[3].toDouble() / 1024.0);
+				// qDebug("挂载点:%s 已用:%.0lfMB 可用:%.0lfMB", lst[5]helper::Qstr2str().c_str(), lst[2].toDouble() / 1024.0, lst[3].toDouble() / 1024.0);
 				lFreeAll += lst[2].toDouble() / 1024.0;
 			lTotalAll += lst[3].toDouble() / 1024.0 + lFreeAll;
 		}
@@ -177,8 +177,8 @@ bool MySysInfo::GetdiskSpace(unsigned long& lFreeAll, unsigned long& lTotalAll) 
 
 	static char path[_MAX_PATH]; //存储当前系统存在的盘符
 	int curdrive = _getdrive();
-	lFreeAll	 = 0UL;
-	lTotalAll	 = 0UL;
+	lFreeAll = 0UL;
+	lTotalAll = 0UL;
 	for (int drive = 1; drive <= curdrive; drive++) //遍历所有盘符
 	{
 		if (!_chdrive(drive)) {
@@ -186,13 +186,13 @@ bool MySysInfo::GetdiskSpace(unsigned long& lFreeAll, unsigned long& lTotalAll) 
 			ULARGE_INTEGER caller, total, free;
 			WCHAR wszClassName[_MAX_PATH];
 			memset(wszClassName, 0, sizeof(wszClassName));
-			MultiByteToWideChar(CP_ACP, 0, path, strlen(path) + 1, wszClassName, sizeof(wszClassName) / sizeof(wszClassName[0]));
+			MultiByteToWideChar(CP_ACP, 0, path, static_cast<int>(strlen(path)) + 1, wszClassName, sizeof(wszClassName) / sizeof(wszClassName[0]));
 			if (GetDiskFreeSpaceEx(wszClassName, &caller, &total, &free) == 0) {
 				// qDebug() << "GetDiskFreeSpaceEx Filed!";
 				return false;
 			}
 
-			double dTepFree	 = free.QuadPart / (1024.0 * 1024.0);
+			double dTepFree = free.QuadPart / (1024.0 * 1024.0);
 			double dTepTotal = total.QuadPart / (1024.0 * 1024.0);
 			// qDebug() << "Get Windows Disk Information:" << path << "--free:" << dTepFree << "MB,--total:" << dTepTotal << "MB";
 			lFreeAll += ceil(dTepFree);
